Node out of LDAP: rebased on Rudder 8.0

+++ webapp/sources/utils/src/main/scala/com/normation/utils/Version.scala
@@ -40,6 +40,8 @@ package com.normation.utils
 import java.nio.charset.Charset
 import zio.Chunk
 
+import zio.Chunk
+
 /**
  * When comparing two version numbers, first the epoch of each are compared, then
  * the upstream_version if epoch is equal. epoch is compared numerically. The
@@ -236,8 +238,8 @@ object ParseVersion {
     case (seq, n) => // seq is at least 1
       seq.last match {
         case Separator.Tilde =>
-          listOfSepToPart(seq.init.toList) ::: VersionPart.Before(Separator.Tilde, PartType.Numeric(n)) :: Nil
-        case sep             => listOfSepToPart(seq.init.toList) ::: VersionPart.After(sep, PartType.Numeric(n)) :: Nil
+          listOfSepToPart(seq.init.toList) ::: VersionPart.Before(Separator.Tilde, PartType.Numeric(n.toInt)) :: Nil
+        case sep             => listOfSepToPart(seq.init.toList) ::: VersionPart.After(sep, PartType.Numeric(n.toInt)) :: Nil
       }
   }
 
@@ -251,7 +253,7 @@ object ParseVersion {
   }
 
   def noSepPart1[A: P] = P(chars).map(c => VersionPart.After(Separator.None, c) :: Nil)
-  def noSepPart2[A: P] = P(num).map(n => VersionPart.After(Separator.None, PartType.Numeric(n)) :: Nil)
+  def noSepPart2[A: P] = P(num).map(n => VersionPart.After(Separator.None, PartType.Numeric(n.toInt)) :: Nil)
 
   def startNum[A: P] = P(num).map(i => PartType.Numeric(i))
